+++ src/by/it/mialeshka/jd02_02/Cashier.java
@@ -10,12 +10,12 @@ class Cashier implements Runnable{
         name = "\tCashier №" + number;
         countServBuyer = 0;
         numberCashier = number;
-        Manager.cashierStart();
+        Dispatcher.cashierStart();
     }
 
     @Override
     public void run() {
-        while (!Manager.planComplate()){
+        while (!Dispatcher.planComplate()){
             waitCashier();
             Buyer buyer = QueueBuyers.extract();
             if (buyer != null){
@@ -43,8 +43,8 @@ class Cashier implements Runnable{
 
     private void waitCashier(){
         synchronized (this){
-            if((Manager.getWorkCashier()*5 > QueueBuyers.size() && QueueBuyers.size() > 0) || Manager.startWorkShop()){
-                Manager.cashierStop();
+            if((Dispatcher.getWorkCashier()*5 > QueueBuyers.size() && QueueBuyers.size() > 0) || Dispatcher.startWorkShop()){
+                Dispatcher.cashierStop();
                 ListCashier.add(this);
                 try {
                     if (this.countServBuyer > 0) System.out.println(this + " closed");
@@ -71,17 +71,17 @@ class Cashier implements Runnable{
                 .append("\n");
 
         for (Map.Entry<String, Double> good: buyer.goods.entrySet()) {
-            for (int i = 1; i <= Manager.PLAN_CASHIER; i++) {
+            for (int i = 1; i <= Dispatcher.PLAN_CASHIER; i++) {
                 if (numberCashier == i){
                     check.append(String.format("%-22s", String.format("%s - %10.2f", good.getKey(), good.getValue())));
-                    Manager.addPrice(good.getValue());
+                    Dispatcher.addPrice(good.getValue());
                 }
                 else{
                     check.append(String.format("%22s", ""));
                 }
             }
             check.append(String.format("%-22d", QueueBuyers.size()));
-            check.append(String.format("%-22.2f", Manager.TOTAL_SUMM)).append("\n");
+            check.append(String.format("%-22.2f", Dispatcher.TOTAL_SUMM)).append("\n");
         }
         return check.toString();
     }
